added id for CarouselInput
